Move RestService into rest module
