+++ gradle.properties
@@ -8,5 +8,6 @@ ktorVersion=1.4.0
 kotlin.code.style=official
 kotlinVersion=1.4.0
 logbackVersion=1.2.1
+slf4jVersion=1.7.30
 
 bmuschkoVersion=6.6.1
