+++ ok-catalogue-transport-common/src/main/kotlin/ru/otus/otuskotlin/catalogue/transport/common/models/ErrorDTO.kt
@@ -1,0 +1,17 @@
+package ru.otus.otuskotlin.catalogue.transport.common.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ErrorDTO(
+        var code:String? = null,
+        var level: Level? = null,
+        var message:String? = null
+) {
+        @Serializable
+        enum class Level{
+            NONE,
+            WARNING,
+            ERROR
+        }
+}
