+++ ok-catalogue-transport-common/src/main/kotlin/ru/otus/otuskotlin/catalogue/transport/common/models/ResponseModel.kt
@@ -1,0 +1,9 @@
+package ru.otus.otuskotlin.catalogue.transport.common.models
+
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.Transient
+
+@Serializable
+abstract class ResponseModel(
+        @Transient open var status: ErrorDTO? = null
+)
