--- ok-catalogue-transport-common/src/main/kotlin/ru/otus/otuskotlin/catalogue/transport/common/models/categories/CategoryError.kt
@@ -1,18 +1,0 @@
-package ru.otus.otuskotlin.catalogue.transport.common.models.categories
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class CategoryError(
-    var code:String? = null,
-    var level: Level? = null,
-    var message:String? = null
-) {
-        @Serializable
-        enum class Level{
-            NONE,
-            SUCCESS,
-            WARNING,
-            ERROR
-        }
-}
