+++ ok-catalogue-transport-common/src/main/kotlin/ru/otus/otuskotlin/catalogue/transport/common/models/categories/CategoryGetMapResponse.kt
@@ -1,11 +1,13 @@
 package ru.otus.otuskotlin.catalogue.transport.common.models.categories
 
 import kotlinx.serialization.Serializable
+import ru.otus.otuskotlin.catalogue.transport.common.models.ErrorDTO
+import ru.otus.otuskotlin.catalogue.transport.common.models.ResponseModel
 
 @Serializable
 data class CategoryGetMapResponse(
     var data:CategoryMapDTO? = null,
-    var status:CategoryError? = null
-) {
-
-}
+    override var status: ErrorDTO? = null
+): ResponseModel(
+        status = status
+)
