+++ ok-catalogue-transport-common/src/main/kotlin/ru/otus/otuskotlin/catalogue/transport/common/models/categories/CategoryGetResponse.kt
@@ -1,7 +1,8 @@
 package ru.otus.otuskotlin.catalogue.transport.common.models.categories
 
 import kotlinx.serialization.Serializable
-import ru.otus.otuskotlin.catalogue.transport.common.models.items.ItemInfo
+import ru.otus.otuskotlin.catalogue.transport.common.models.ErrorDTO
+import ru.otus.otuskotlin.catalogue.transport.common.models.ResponseModel
 
 /**
  *  Model for all responses by categories queries
@@ -9,6 +10,7 @@ import ru.otus.otuskotlin.catalogue.transport.common.models.items.ItemInfo
 @Serializable
 data class CategoryGetResponse(
         var data: CategoryDTO? = null,
-        var status: CategoryError? = null
-) {
-}
+        override var status: ErrorDTO? = null
+): ResponseModel(
+        status = status
+)
