+++ ok-catalogue-transport-common/src/main/kotlin/ru/otus/otuskotlin/catalogue/transport/common/models/categories/CategoryResponse.kt
@@ -2,8 +2,7 @@ package ru.otus.otuskotlin.catalogue.transport.common.models.categories
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
-import ru.otus.otuskotlin.catalogue.transport.common.models.categories.CategoryError
-import ru.otus.otuskotlin.catalogue.transport.common.models.categories.CategoryInfo
+import ru.otus.otuskotlin.catalogue.transport.common.models.ErrorDTO
 
 /**
  *  Deprecated
@@ -14,6 +13,6 @@ abstract class CategoryResponse(
         @Transient open var parentList:MutableList<CategoryInfo>? = null,
         @Transient open var type:String? = null,
         @Transient open var label:String? = null,
-        @Transient open var status: CategoryError? = null
+        @Transient open var status: ErrorDTO? = null
 ) {
 }
