+++ ok-catalogue-transport-common/src/main/kotlin/ru/otus/otuskotlin/catalogue/transport/common/models/items/ItemResponse.kt
@@ -1,11 +1,13 @@
 package ru.otus.otuskotlin.catalogue.transport.common.models.items
 
 import kotlinx.serialization.Serializable
-import ru.otus.otuskotlin.catalogue.transport.common.models.categories.CategoryError
+import ru.otus.otuskotlin.catalogue.transport.common.models.ErrorDTO
+import ru.otus.otuskotlin.catalogue.transport.common.models.ResponseModel
 
 @Serializable
 data class ItemResponse(
     var data: ItemInfo? = null,
-    var status: CategoryError? = null
-) {
-}
+    override var status: ErrorDTO? = null
+): ResponseModel(
+        status = status
+)
