+++ ok-catalogue-transport-rest/build.gradle.kts
@@ -1,5 +1,8 @@
+val logbackVersion: String by project
+
 plugins {
     kotlin("jvm")
+
 }
 
 group = rootProject.group
@@ -10,6 +13,10 @@ repositories {
 }
 
 dependencies {
+    val slf4jVersion: String by project
+
+    implementation("org.slf4j:slf4j-api:$slf4jVersion")
+
     implementation(kotlin("stdlib"))
     implementation(project(":ok-catalogue-backend-common"))
     implementation(project(":ok-catalogue-transport-common"))
