+++ ok-catalogue-transport-rest/src/main/kotlin/ru/otus/otuskotlin/catalogue/transport/rest/MainMappers.kt
@@ -2,15 +2,12 @@ package ru.otus.otuskotlin.catalogue.transport.rest
 
 import ru.otus.otuskotlin.catalogue.backend.common.CategoryContext
 import ru.otus.otuskotlin.catalogue.backend.common.CategoryContextStatus
-import ru.otus.otuskotlin.catalogue.backend.common.models.CategoryInfoModel
 import ru.otus.otuskotlin.catalogue.backend.common.models.CategoryModel
-import ru.otus.otuskotlin.catalogue.backend.common.models.CategoryType
+import ru.otus.otuskotlin.catalogue.transport.common.models.ErrorDTO
+import ru.otus.otuskotlin.catalogue.transport.common.models.ResponseModel
 import ru.otus.otuskotlin.catalogue.transport.common.models.categories.*
-import ru.otus.otuskotlin.catalogue.transport.common.models.items.ItemDeleteQuery
-import ru.otus.otuskotlin.catalogue.transport.common.models.items.ItemInfo
 import ru.otus.otuskotlin.catalogue.transport.common.models.items.ItemResponse
 import java.lang.ClassCastException
-import java.lang.IllegalArgumentException
 import java.time.LocalDate
 import java.util.*
 
@@ -113,13 +110,13 @@ fun CategoryModel.toTreeDTO():CategoryMapDTO {
 //    )
 //}
 
-fun CategoryContextStatus.toDTO() = CategoryError(
-        level = CategoryError.Level.valueOf(this.toString())
+fun CategoryContextStatus.toDTO() = ErrorDTO(
+        level = ErrorDTO.Level.valueOf(this.toString())
 )
 
 internal fun String.toDTOString() = this.takeIf { it.isNotBlank() }
 
-inline fun <reified T> CategoryContext.getResult():T{
+inline fun <reified T: ResponseModel> CategoryContext.getResult():T{
     return when(T::class){
         CategoryGetResponse::class -> resultCategory() as T
         ItemResponse::class -> resultItem() as T
