+++ ok-catalogue-transport-rest/src/test/kotlin/ru/otus/otuskotlin/catalogue/transport/rest/MainMappersKtTest.kt
@@ -2,12 +2,10 @@ package ru.otus.otuskotlin.catalogue.transport.rest
 
 import ru.otus.otuskotlin.catalogue.backend.common.CategoryContext
 import ru.otus.otuskotlin.catalogue.backend.common.CategoryContextStatus
-import ru.otus.otuskotlin.catalogue.backend.common.models.CategoryInfoModel
 import ru.otus.otuskotlin.catalogue.backend.common.models.CategoryModel
 import ru.otus.otuskotlin.catalogue.backend.common.models.CategoryType
 import ru.otus.otuskotlin.catalogue.backend.common.models.items.NoteModel
-import ru.otus.otuskotlin.catalogue.transport.common.models.categories.CategoryCreateQuery
-import ru.otus.otuskotlin.catalogue.transport.common.models.categories.CategoryError
+import ru.otus.otuskotlin.catalogue.transport.common.models.ErrorDTO
 import ru.otus.otuskotlin.catalogue.transport.common.models.items.ItemInfo
 import ru.otus.otuskotlin.catalogue.transport.common.models.items.NoteInfo
 import kotlin.test.Test
@@ -33,7 +31,7 @@ internal class MainMappersKtTest{
         var context = CategoryContext()
         context.status = CategoryContextStatus.WARNING
         var error = context.status.toDTO()
-        assertEquals(CategoryError.Level.WARNING, error.level)
+        assertEquals(ErrorDTO.Level.WARNING, error.level)
     }
 
     @Test
